refactor: rename file companies form.py / paginate
