+++ apps/companies/forms/__init__.py
@@ -1,1 +1,1 @@
-from .company_form import CompanyForm
+from .companies_form import CompanyForm
